refactor(drag-sort): separate drag logic from UI in DragSortClient
Extract pointer/drag-and-drop state into a useDragSortDrag hook and
split the menu/result/play screens into their own components, so
DragSortClient is now a thin orchestrator per the Style D convention.
